@@ -1,16 +1,14 @@
 name: google_language_fonts_tester
 description: A Flutter application showcasing how to use the google_language_fonts package.
-
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.2
   google_language_fonts:
     path: ../
 
@@ -20,6 +18,5 @@ dev_dependencies:
 
 flutter:
   uses-material-design: true
-
   assets:
     - google_fonts/
