@@ -1,14 +1,11 @@
-import 'package:equatable/equatable.dart';
-import 'package:meta/meta.dart';
-
 /// Used for fetching from [google-webfonts-helper](https://google-webfonts-helper.herokuapp.com/).
-class Font extends Equatable {
+class Font {
   final String family;
   final List<String> langSubsets;
 
   const Font({
-    @required this.family,
-    @required this.langSubsets,
+    required this.family,
+    required this.langSubsets,
   });
 
   factory Font.fromJson(Map<String, dynamic> json) => Font(
@@ -22,17 +19,25 @@ class Font extends Equatable {
       };
 
   @override
-  List<Object> get props => [family, langSubsets];
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is Font &&
+          runtimeType == other.runtimeType &&
+          family == other.family &&
+          langSubsets == other.langSubsets;
+
+  @override
+  int get hashCode => family.hashCode ^ langSubsets.hashCode;
 }
 
 /// This model class is used in errors.json file and contains font names and
 /// error phrases of unrecognized fonts while fetching from official Google Fonts API.
 class UnrecognizedFont {
   final String fontName;
-  final String errorPhrase;
+  final String? errorPhrase;
 
   const UnrecognizedFont({
-    @required this.fontName,
+    required this.fontName,
     this.errorPhrase,
   });
 
