@@ -41,31 +41,28 @@ void clearCache() => _loadedFonts.clear();
 /// This function has a side effect of loading the font into the [FontLoader],
 /// either by network or from the device file system.
 TextStyle googleFontsTextStyle({
-  @required String fontFamily,
-  TextStyle textStyle,
-  Color color,
-  Color backgroundColor,
-  double fontSize,
-  FontWeight fontWeight,
-  FontStyle fontStyle,
-  double letterSpacing,
-  double wordSpacing,
-  TextBaseline textBaseline,
-  double height,
-  Locale locale,
-  Paint foreground,
-  Paint background,
-  List<Shadow> shadows,
-  List<FontFeature> fontFeatures,
-  TextDecoration decoration,
-  Color decorationColor,
-  TextDecorationStyle decorationStyle,
-  double decorationThickness,
-  @required Map<GoogleFontsVariant, GoogleFontsFile> fonts,
+  required String fontFamily,
+  TextStyle? textStyle,
+  Color? color,
+  Color? backgroundColor,
+  double? fontSize,
+  FontWeight? fontWeight,
+  FontStyle? fontStyle,
+  double? letterSpacing,
+  double? wordSpacing,
+  TextBaseline? textBaseline,
+  double? height,
+  Locale? locale,
+  Paint? foreground,
+  Paint? background,
+  List<Shadow>? shadows,
+  List<FontFeature>? fontFeatures,
+  TextDecoration? decoration,
+  Color? decorationColor,
+  TextDecorationStyle? decorationStyle,
+  double? decorationThickness,
+  required Map<GoogleFontsVariant, GoogleFontsFile> fonts,
 }) {
-  assert(fontFamily != null);
-  assert(fonts != null);
-
   textStyle ??= TextStyle();
   textStyle = textStyle.copyWith(
     color: color,
@@ -100,7 +97,7 @@ TextStyle googleFontsTextStyle({
 
   final descriptor = GoogleFontsDescriptor(
     familyWithVariant: familyWithVariant,
-    file: fonts[matchedVariant],
+    file: fonts[matchedVariant]!,
   );
 
   loadFontIfNecessary(descriptor);
@@ -134,7 +131,7 @@ Future<void> loadFontIfNecessary(GoogleFontsDescriptor descriptor) async {
   }
 
   try {
-    Future<ByteData> byteData;
+    Future<ByteData?>? byteData;
 
     // Check if this font can be loaded by the pre-bundled assets.
     final assetManifestJson = await assetManifest.json();
@@ -146,14 +143,14 @@ Future<void> loadFontIfNecessary(GoogleFontsDescriptor descriptor) async {
       byteData = rootBundle.load(assetPath);
     }
     if (await byteData != null) {
-      return _loadFontByteData(familyWithVariantString, byteData);
+      return loadFontByteData(familyWithVariantString, byteData);
     }
 
     // Check if this font can be loaded from the device file system.
     byteData = file_io.loadFontFromDeviceFileSystem(familyWithVariantString);
 
     if (await byteData != null) {
-      return _loadFontByteData(familyWithVariantString, byteData);
+      return loadFontByteData(familyWithVariantString, byteData);
     }
 
     // Attempt to load this font via http, unless disallowed.
@@ -163,13 +160,13 @@ Future<void> loadFontIfNecessary(GoogleFontsDescriptor descriptor) async {
         descriptor.file,
       );
       if (await byteData != null) {
-        return _loadFontByteData(familyWithVariantString, byteData);
+        return loadFontByteData(familyWithVariantString, byteData);
       }
     } else {
       throw Exception(
-        "GoogleFonts.config.allowRuntimeFetching is false but font $fontName was not "
-        "found in the application assets. Ensure $fontName.otf exists in a "
-        "folder that is included in your pubspec's assets.",
+        'GoogleFonts.config.allowRuntimeFetching is false but font $fontName was not '
+        'found in the application assets. Ensure $fontName.otf exists in a '
+        'folder that is included in your pubspec\'s assets.',
       );
     }
   } catch (e) {
@@ -181,16 +178,18 @@ Future<void> loadFontIfNecessary(GoogleFontsDescriptor descriptor) async {
 }
 
 /// Loads a font with [FontLoader], given its name and byte-representation.
-Future<void> _loadFontByteData(
+@visibleForTesting
+Future<void> loadFontByteData(
   String familyWithVariantString,
-  Future<ByteData> byteData,
+  Future<ByteData?>? byteData,
 ) async {
-  final anyFontDataFound = byteData != null && await byteData != null;
-  if (anyFontDataFound) {
-    final fontLoader = FontLoader(familyWithVariantString);
-    fontLoader.addFont(byteData);
-    await fontLoader.load();
-  }
+  if (byteData == null) return;
+  final fontData = await byteData;
+  if (fontData == null) return;
+
+  final fontLoader = FontLoader(familyWithVariantString);
+  fontLoader.addFont(Future.value(fontData));
+  await fontLoader.load();
 }
 
 /// Returns [GoogleFontsVariant] from [variantsToCompare] that most closely
@@ -203,8 +202,8 @@ GoogleFontsVariant _closestMatch(
   GoogleFontsVariant sourceVariant,
   Iterable<GoogleFontsVariant> variantsToCompare,
 ) {
-  int bestScore;
-  GoogleFontsVariant bestMatch;
+  int? bestScore;
+  late GoogleFontsVariant bestMatch;
   for (final variantToCompare in variantsToCompare) {
     final score = _computeMatch(sourceVariant, variantToCompare);
     if (bestScore == null || score < bestScore) {
@@ -260,7 +259,7 @@ int _computeMatch(GoogleFontsVariant a, GoogleFontsVariant b) {
   if (a == b) {
     return 0;
   }
-  int score = (a.fontWeight.index - b.fontWeight.index).abs();
+  var score = (a.fontWeight.index - b.fontWeight.index).abs();
   if (a.fontStyle != b.fontStyle) {
     score += 2;
   }
@@ -269,16 +268,16 @@ int _computeMatch(GoogleFontsVariant a, GoogleFontsVariant b) {
 
 /// Looks for a matching [familyWithVariant] font, provided the asset manifest.
 /// Returns the path of the font asset if found, otherwise an empty string.
-String _findFamilyWithVariantAssetPath(
+String? _findFamilyWithVariantAssetPath(
   GoogleFontsFamilyWithVariant familyWithVariant,
-  Map<String, List<String>> manifestJson,
+  Map<String, List<String>>? manifestJson,
 ) {
   if (manifestJson == null) return null;
 
   final apiFilenamePrefix = familyWithVariant.toApiFilenamePrefix();
 
   for (final assetList in manifestJson.values) {
-    for (final String asset in assetList) {
+    for (final asset in assetList) {
       for (final matchingSuffix in ['.ttf', '.otf'].where(asset.endsWith)) {
         final assetWithoutExtension =
             asset.substring(0, asset.length - matchingSuffix.length);
