@@ -4,15 +4,12 @@
 
 import 'dart:ui';
 
-import 'package:flutter/foundation.dart';
-
 /// Represents a Google Fonts API variant in Flutter-specific types.
 class GoogleFontsVariant {
   const GoogleFontsVariant({
-    @required this.fontWeight,
-    @required this.fontStyle,
-  })  : assert(fontWeight != null),
-        assert(fontStyle != null);
+    required this.fontWeight,
+    required this.fontStyle,
+  });
 
   /// Creates a [GoogleFontsVariant] from a Google Fonts API specific
   /// filename part.
@@ -29,8 +26,8 @@ class GoogleFontsVariant {
   ///
   /// See [GoogleFontsVariant.toApiFilenamePart] for the inverse function.
   GoogleFontsVariant.fromApiFilenamePart(String filenamePart)
-      : this.fontWeight = _extractFontWeightFromApiFilenamePart(filenamePart),
-        this.fontStyle = _extractFontStyleFromApiFilenamePart(filenamePart);
+      : fontWeight = _extractFontWeightFromApiFilenamePart(filenamePart),
+        fontStyle = _extractFontStyleFromApiFilenamePart(filenamePart);
 
   /// Creates a [GoogleFontsVariant] from a Google Fonts API specific
   /// variant name.
@@ -43,11 +40,11 @@ class GoogleFontsVariant {
   ///
   /// See [GoogleFontsVariant.toString] for the inverse function.
   GoogleFontsVariant.fromString(String variantString)
-      : this.fontWeight = FontWeight.values[variantString == _regular ||
+      : fontWeight = FontWeight.values[variantString == _regular ||
                 variantString == _italic
             ? 3
             : (int.parse(variantString.replaceAll(_italic, '')) ~/ 100) - 1],
-        this.fontStyle = variantString.contains(_italic)
+        fontStyle = variantString.contains(_italic)
             ? FontStyle.italic
             : FontStyle.normal;
 
@@ -94,7 +91,7 @@ class GoogleFontsVariant {
   /// See [GoogleFontsVariant.fromApiFilenamePart] for the inverse function.
   String toApiFilenamePart() {
     final weightPrefix = _fontWeightToFilenameWeightParts[fontWeight] ??
-        _fontWeightToFilenameWeightParts[FontWeight.w400];
+        _fontWeightToFilenameWeightParts[FontWeight.w400]!;
     final italicSuffix = fontStyle == FontStyle.italic ? 'Italic' : '';
     if (weightPrefix == 'Regular') {
       return italicSuffix == '' ? weightPrefix : italicSuffix;
